backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database and ORM
sqlmodel==0.0.14
sqlalchemy==2.0.23

# Authentication and security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6

# Environment variables
python-dotenv==1.0.0

# Email validation
email-validator==2.1.0

# CORS middleware (included in FastAPI)
# Development dependencies (optional)
# pytest==7.4.3
# httpx==0.25.2